Split bridge frame result metrics
